Enable 관련 세팅

+++ SITemplate.Setting/Converters/WindowThemeValueConverter.cs
@@ -3,9 +3,6 @@ using CvsService.UI.Windows.Enums;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Data;
 
 namespace SITemplate.Setting.Converters
@@ -28,7 +25,7 @@ namespace SITemplate.Setting.Converters
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return (EWindowTheme)value;
+            return (EWindowTheme)(int)value;
         }
 
         public static string GetString(EWindowTheme value)
